requirements.txt
fastapi==0.115.5
firebase_admin==6.6.0
libcst==1.5.1
pydantic_settings==2.6.1
python-dotenv==1.0.1
